refactor Preferences
Encapsulated path Locations in to a Locations class and Operating System
specific information in to an OperatingSystem enum.

The enum includes path info specific to each OS that the Locations class
uses to build up the correct paths.

Preferences has been refactored to make use of the Locations class.

Additionally, the version is now only maintained in the build.gradle
file. It is built in to the jar's MANIFEST.MF with the key
'Application-Version'. This is grabbed at runtime to determine the
version. Of course, when ran through an IDE there is no MANIFEST.MF file
so the version defaults to "Debug".

Because of these changes, the build is currently broken until the
Updater is fixed and handling settings in the developer case is handled.