Added ability to see what's going on behind overlay
Now when you bring up the overlay you can see what is happening behind
it. When you drag a rectangle it will display the initial overlay
screenshot in place covering the background movement. If you middle
click the mouse wheel, it will take a bran new screenshot. If you scroll
the mouse wheel up or down it will change the transparency of the
dragged sub image and the surrounding overlay allowing you to see
movement behind your subimage and overlay.